% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geneApp.R
\name{run_geneApp}
\alias{run_geneApp}
\title{geneApp launcher}
\usage{
run_geneApp()
}
\value{
The running application
}
\description{
This is the launcher function to the geneApp package, The only way to first start the app
}
