% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneApp-package.R
\docType{package}
\name{geneApp-package}
\alias{geneApp}
\alias{geneApp-package}
\title{geneApp: geneApp - Exploratory data analysis in genetic studies}
\description{
Package containing geneApp, an application for exploratory data analysis in genetic studies. All the app is self contained, you should launch it using run_geneApp().
}
\author{
\strong{Maintainer}: Filippo Mazzara \email{mazzara.1742740@studenti.uniroma1.it}

}
\keyword{internal}
